model/requirements.txt
numpy==2.2.2
pillow==11.1.0
tqdm==4.66.6
huggingface-hub==0.27.1
matplotlib==3.10.0

torch==2.5.1+cu124
torchvision==0.20.1+cu124
torchview==0.2.6
torchviz==0.0.3

fastapi==0.115.6
pydantic==2.10.4
pydantic_core==2.27.2